Split common/test/symbolic_expression.test.cc file
Close https://github.com/RobotLocomotion/drake/issues/4848
