+++ drake/common/test/symbolic_expression_matrix_test.cc
@@ -1,0 +1,134 @@
+#include "drake/common/symbolic_expression.h"
+
+#include "gtest/gtest.h"
+
+#include "drake/common/symbolic_variable.h"
+
+namespace drake {
+namespace symbolic {
+namespace {
+
+class SymbolicExpressionMatrixTest : public ::testing::Test {
+ protected:
+  const Variable var_x_{"x"};
+  const Variable var_y_{"y"};
+  const Variable var_z_{"z"};
+  const Expression x_{var_x_};
+  const Expression y_{var_y_};
+  const Expression z_{var_z_};
+
+  const Expression zero_{0.0};
+  const Expression one_{1.0};
+  const Expression two_{2.0};
+  const Expression neg_one_{-1.0};
+  const Expression pi_{3.141592};
+  const Expression neg_pi_{-3.141592};
+  const Expression e_{2.718};
+
+  Eigen::Matrix<Expression, 3, 2, Eigen::DontAlign> A_;
+  Eigen::Matrix<Expression, 2, 3, Eigen::DontAlign> B_;
+  Eigen::Matrix<Expression, 3, 2, Eigen::DontAlign> C_;
+
+  void SetUp() override {
+    // clang-format off
+    A_ << x_, one_,       //  [x  1]
+          y_, neg_one_,   //  [y -1]
+          z_, pi_;        //  [z  3.141592]
+
+    B_ << x_, y_,  z_,    //  [x     y        z]
+          e_, pi_, two_;  //  [2.718 3.141592 2]
+
+    C_ << z_, two_,       //  [z  2]
+          x_, e_,         //  [x -2.718]
+          y_, pi_;        //  [y  3.141592]
+    // clang-format on
+  }
+};
+
+TEST_F(SymbolicExpressionMatrixTest, EigenAdd) {
+  auto const M(A_ + A_);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  // clang-format off
+  M_expected << (x_ + x_), (one_ + one_),
+                (y_ + y_), (neg_one_ + neg_one_),
+                (z_ + z_), (pi_ + pi_);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenSub1) {
+  auto const M(A_ - A_);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  EXPECT_EQ(M, M_expected);  // should be all zero.
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenSub2) {
+  auto const M(A_ - C_);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  // clang-format off
+  M_expected << (x_ - z_), (one_ - two_),
+                (y_ - x_), (neg_one_ - e_),
+                (z_ - y_), (pi_ - pi_);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);  // should be all zero.
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenMul1) {
+  auto const M(A_ * B_);
+  Eigen::Matrix<Expression, 3, 3> M_expected;
+  // clang-format off
+  M_expected <<
+    (x_ * x_ + e_),       (x_ * y_ + pi_),       (x_ * z_ + two_),
+    (y_ * x_ + -e_),      (y_ * y_ + - pi_),     (y_ * z_ + - two_),
+    (z_ * x_ + pi_ * e_), (z_ * y_ + pi_ * pi_), (z_ * z_ + pi_ * two_);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenMul2) {
+  auto const M(B_ * A_);
+  Eigen::Matrix<Expression, 2, 2> M_expected;
+  // clang-format off
+  M_expected <<
+    (x_ * x_ + (y_ * y_ + z_ * z_)),    (x_ + (-y_ + z_ * pi_)),
+    (e_ * x_ + (pi_ * y_ + two_ * z_)), (e_ * one_ + pi_ * - one_ + two_ * pi_);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenMul3) {
+  auto const M(2.0 * A_);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  // clang-format off
+  M_expected << (2 * x_), (2 * one_),
+                (2 * y_), (2 * neg_one_),
+                (2 * z_), (2 * pi_);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenMul4) {
+  auto const M(A_ * 2.0);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  // clang-format off
+  M_expected << (x_ * 2), (one_ * 2),
+                (y_ * 2), (neg_one_ * 2),
+                (z_ * 2), (pi_ * 2);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+TEST_F(SymbolicExpressionMatrixTest, EigenDiv) {
+  auto const M(A_ / 2.0);
+  Eigen::Matrix<Expression, 3, 2> M_expected;
+  // clang-format off
+  M_expected << (x_ / 2), (one_ / 2),
+                (y_ / 2), (neg_one_ / 2),
+                (z_ / 2), (pi_ / 2);
+  // clang-format on
+  EXPECT_EQ(M, M_expected);
+}
+
+}  // namespace
+}  // namespace symbolic
+}  // namespace drake
